whoops a lot of changes at once
PixelFormat stuff into its own module (incomplete)
Palette module
Surface gets clip rect support and palette support
